问诊列表

+++ biz_case/src/main/java/com/ruoyi/bizcase/service/IBizConsultationSessionsService.java
@@ -1,19 +1,19 @@
 package com.ruoyi.bizcase.service;
 
 import java.util.List;
+
 import com.ruoyi.bizcase.domain.BizConsultationSessions;
 
 /**
  * 学生问诊列Service接口
- * 
+ *
  * @author daizor
  * @date 2025-09-04
  */
-public interface IBizConsultationSessionsService 
-{
+public interface IBizConsultationSessionsService {
     /**
      * 查询学生问诊列
-     * 
+     *
      * @param sessionId 学生问诊列主键
      * @return 学生问诊列
      */
@@ -21,7 +21,7 @@ public interface IBizConsultationSessionsService
 
     /**
      * 查询学生问诊列列表
-     * 
+     *
      * @param bizConsultationSessions 学生问诊列
      * @return 学生问诊列集合
      */
@@ -29,15 +29,15 @@ public interface IBizConsultationSessionsService
 
     /**
      * 新增学生问诊列
-     * 
+     *
      * @param bizConsultationSessions 学生问诊列
-     * @return 结果
+     * @return 新增记录的sessionId
      */
-    public int insertBizConsultationSessions(BizConsultationSessions bizConsultationSessions);
+    public Long insertBizConsultationSessions(BizConsultationSessions bizConsultationSessions);
 
     /**
      * 修改学生问诊列
-     * 
+     *
      * @param bizConsultationSessions 学生问诊列
      * @return 结果
      */
@@ -45,7 +45,7 @@ public interface IBizConsultationSessionsService
 
     /**
      * 批量删除学生问诊列
-     * 
+     *
      * @param sessionIds 需要删除的学生问诊列主键集合
      * @return 结果
      */
@@ -53,7 +53,7 @@ public interface IBizConsultationSessionsService
 
     /**
      * 删除学生问诊列信息
-     * 
+     *
      * @param sessionId 学生问诊列主键
      * @return 结果
      */
